Added panellhnies reminder and clear command.

+++ src/main/java/me/acrispycookie/commands/MessageCommand.java
@@ -20,17 +20,25 @@ public class MessageCommand extends BotCommand {
     @Override
     public void execute(String[] args, String label, Member m, TextChannel t, List<Member> mentions, List<Role> mentionedRoles, List<Message.Attachment> attachments, Message message) {
         if(Perm.hasPermission(m, Perm.MESSAGE)){
-            message.delete().queue();
-            if(Utils.isInt(args[0]) && m.getGuild().getTextChannelById(args[0]) != null){
-                TextChannel textChannel = m.getGuild().getTextChannelById(args[0]);
-                ArrayList<String> newArgs = new ArrayList<>(Arrays.asList(args));
-                newArgs.remove(0);
-                String msg = newArgs.stream().map(Objects::toString).collect(Collectors.joining(" "));
-                textChannel.sendMessage(msg).queue();
+            if(args.length > 0){
+                message.delete().queue();
+                if(Utils.isInt(args[0]) && m.getGuild().getTextChannelById(args[0]) != null){
+                    TextChannel textChannel = m.getGuild().getTextChannelById(args[0]);
+                    ArrayList<String> newArgs = new ArrayList<>(Arrays.asList(args));
+                    newArgs.remove(0);
+                    String msg = newArgs.stream().map(Objects::toString).collect(Collectors.joining(" "));
+                    textChannel.sendMessage(msg).queue();
+                }
+                else{
+                    String msg = Arrays.stream(args).map(Objects::toString).collect(Collectors.joining(" "));
+                    t.sendMessage(msg).queue();
+                }
             }
             else{
-                String msg = Arrays.stream(args).map(Objects::toString).collect(Collectors.joining(" "));
-                t.sendMessage(msg).queue();
+                t.sendMessage(new EmbedMessage(m.getUser(),
+                        Main.getInstance().getLanguageManager().get("commands.invalid.title"),
+                        Main.getInstance().getLanguageManager().get("commands.invalid.description.message"),
+                        Main.getInstance().getErrorColor()).build()).queue();
             }
         }
         else{
